Commands are stoppable now, no more "wrong thread" errors
