+++ src/main/java/org/asciidoc/intellij/commandRunner/arbitrary/AsciiDocRunnerForPython.java
@@ -1,4 +1,4 @@
-package org.asciidoc.intellij.commandRunner;
+package org.asciidoc.intellij.commandRunner.arbitrary;
 
 import com.intellij.execution.configurations.PathEnvironmentVariableUtil;
 import com.intellij.lang.Language;
@@ -18,24 +18,25 @@ import java.nio.file.Paths;
 import java.util.List;
 import java.util.Locale;
 
-public class AsciiDocRunnerForRuby extends AsciiDocRunnerArbitrary {
-  private static final String RUBY_LANGUAGE_ID = "ruby";
+public class AsciiDocRunnerForPython extends AsciiDocRunnerArbitrary {
+  private static final String PYTHON_LANGUAGE_ID = "python";
 
-  private static final String RUBY_EXE = "ruby";
+  private static final String PYTHON_EXE = "python";
+  private static final String PYTHON3_EXE = "python3";
 
   private static String interpreterCache = null;
 
   @Override
   public boolean isApplicable(@NotNull Project project, Language language) {
     return language != null &&
-      language.getID().equalsIgnoreCase(RUBY_LANGUAGE_ID)
+      language.getID().equalsIgnoreCase(PYTHON_LANGUAGE_ID)
       && findInterpreter(project) != null;
   }
 
   @NonNull
   @Override
   String codeRunParameter() {
-    return "-e";
+    return "-c";
   }
 
   @Nullable
@@ -44,16 +45,16 @@ public class AsciiDocRunnerForRuby extends AsciiDocRunnerArbitrary {
     if (interpreterCache != null) {
       return interpreterCache;
     }
-    // This will most likely only work in RubyMine.
+    // This will most likely only work in PyCharm.
     Sdk sdk = ProjectRootManager.getInstance(project).getProjectSdk();
-    if (sdk != null && isLikelyRubyInterpreter(sdk.getHomePath())) {
+    if (sdk != null && isLikelyPythonInterpreter(sdk.getHomePath())) {
       return sdk.getHomePath();
     }
 
     // This should also work in IntelliJ.
     List<String> candidates = SystemInfo.isWindows
-      ? List.of(RUBY_EXE + ".exe", RUBY_EXE)
-      : List.of(RUBY_EXE);
+      ? List.of("py", PYTHON_EXE, PYTHON3_EXE)
+      : List.of(PYTHON3_EXE, PYTHON_EXE);
 
     for (String candidate : candidates) {
       if (canExecute(candidate)) {
@@ -65,13 +66,13 @@ public class AsciiDocRunnerForRuby extends AsciiDocRunnerArbitrary {
     return null;
   }
 
-  private static boolean isLikelyRubyInterpreter(@Nullable String homePath) {
+  private static boolean isLikelyPythonInterpreter(@Nullable String homePath) {
     if (homePath == null || homePath.isBlank()) {
       return false;
     }
 
     String fileName = Paths.get(homePath).getFileName().toString().toLowerCase(Locale.ROOT);
-    return fileName.startsWith(RUBY_LANGUAGE_ID) || fileName.equals("ruby.exe");
+    return fileName.startsWith(PYTHON_LANGUAGE_ID) || fileName.equals("py") || fileName.equals("py.exe");
   }
 
   private static boolean canExecute(String candidate) {
@@ -88,6 +89,6 @@ public class AsciiDocRunnerForRuby extends AsciiDocRunnerArbitrary {
   @Nls
   @Override
   public String getTitle() {
-    return AsciiDocBundle.message("asciidoc.runner.ruby");
+    return AsciiDocBundle.message("asciidoc.runner.python");
   }
 }
