+++ src/main/java/org/asciidoc/intellij/commandRunner/arbitrary/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * {@link org.asciidoc.intellij.commandRunner.arbitrary.AsciiDocRunnerArbitrary}
+ * is the base-class to run commands / scripts in arbitrary languages.<br>
+ * They are run in the background, so even long-running scripts don't block the UI.
+ */
+package org.asciidoc.intellij.commandRunner.arbitrary;
